bots_n_bolts/born2codebot_dir/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
pycparser==2.19
pylint==2.3.1
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
wrapt==1.11.1
